+++ src/app/utils/cluster-storage/webdav-fields.js
@@ -1,3 +1,5 @@
+const posixRegexp = /^[0-7][0-7][0-7][0-7]?$/;
+
 export default [{
     name: 'endpoint',
     type: 'text',
@@ -20,11 +22,24 @@ export default [{
       { value: 'none', label: 'none' },
       { value: 'basic', label: 'basic' },
       { value: 'token', label: 'token' },
+      { value: 'oauth2', label: 'OAuth2' },
     ],
     tip: true,
   },
   {
     name: 'credentials',
+    type: 'text',
+    tip: true,
+    optional: true,
+  },
+  {
+    name: 'oauth2IdP',
+    type: 'text',
+    tip: true,
+    optional: true,
+  },
+  {
+    name: 'onedataAccessToken',
     type: 'text',
     tip: true,
     optional: true,
@@ -62,16 +77,35 @@ export default [{
     gte: 0,
     tip: true,
     example: 1048576,
-  }, {
+  },
+  {
+    name: 'fileMode',
+    type: 'text',
+    regex: posixRegexp,
+    regexMessage: 'This field should be octal POSIX permissions',
+    example: '0644',
+    optional: true,
+  },
+  {
+    name: 'dirMode',
+    type: 'text',
+    regex: posixRegexp,
+    regexMessage: 'This field should be octal POSIX permissions',
+    example: '0775',
+    optional: true,
+  },
+  {
     name: 'timeout',
     type: 'number',
     optional: true,
-  }, {
+  },
+  {
     name: 'insecure',
     type: 'checkbox',
     defaultValue: false,
     tip: true,
-  }, {
+  },
+  {
     name: 'readonly',
     type: 'checkbox',
     defaultValue: false,