+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-5301 Added new WebDAV configuration fields
 * VFS-5242 Using new async backend to show progress of saving file popularity settings
 * VFS-5187 Fixed some graphical issues in supported spaces view; onepanel client library update
 * VFS-5114 Infinite-scroll list of auto-cleaning reports, file-popularity options and major refactor of space support views